requirements.txt
Pillow ~= 10.0.0
requests ~= 2.31.0
bdfparser ~= 2.2.0